Simplify recursion in photo programs
- Both Photo-Backup and Photo-Renamer used an odd method for recursion.
- Find-CRCs now reports the number of CRCs that were found.
- Photo-Backup uses ReadCrcs and recursive_mkdir from PyBackup.

Photo-Backup needs more testing.

+++ Photo-Backup.py
@@ -1,6 +1,6 @@
 #-------------------------------------------------------------------------------
 # Name: Photo-Backup
-# Purpose: Backup photos based upon the date of those photos
+# Purpose: Backup photos and videos based upon the date they were recorded
 #
 # Author: John Eichenberger
 #
@@ -13,12 +13,10 @@ archive_pictures = "\\Pictures" # the default destination for photo archives
 picture_exts = ['.jpg', '.jpeg',  '.heic',  '.png']
 
 archive_videos = "\\Movies" # the default destination for movie archives
-movie_exts = ['.avi',  '.mpg',  '.mp3',  '.mp4', '.mov',  '.3gp']
+video_exts = ['.avi',  '.mpg',  '.mp3',  '.mp4', '.mov',  '.3gp']
 
 months = [ "01Jan", "02Feb", "03Mar", "04Apr", "05May", "06Jun",
             "07Jul", "08Aug", "09Sep", "10Oct", "11Nov", "12Dec" ]
-
-crcs = {}       # a dictionary of previously backed up files
 
 found = 0       # Count files found
 archived = 0    # Count files archived
@@ -28,94 +26,64 @@ filespec = "*"  # which files get archived?
 
 import glob, re, os, sys
 from crc32 import crc32
-
 from EXIF_Dating import GetExifDate, GetFileDate, GetExifDimensions
+from PyBackup import ReadCrcs, recursive_mkdir
 
-def read_crcs(pn):
-    global crcs
+def remove_duplicate(src, dest):
+    """ Remove a source file that does not need to be backed up """
+    global archived, removed, not_archived
+    try:
+        os.remove(src)
+        print("{} removed as a duplicate of {}".format(src, dest))
+        removed += 1
+    except:
+        print("exception: {} cannot be removed ".format(src))
+        not_archived += 1
 
-    # Read crcs.csv in the specified folder when it is present
-    pn = os.path.expandvars(pn+"\\crcs.csv")
-
-    # Read the CRC file to create a dictionary
-    if not os.path.isfile(pn):
+def replace_destination(src, dest, reason):
+    """ Replace a destination file with a source file for a specified reason """
+    global archived, not_archived
+    try:
+        os.remove(dest)
+    except:
+        print("{}: cannot remove ".format(dest))
+        not_archived += 1
         return
-    f = open(pn)
-    next(f) # skip header line
-    for line in f:
-        sline = line.split(',')
-        crc = sline[0]
-        sline = line.split('"')
-        fn = sline[1]
-        pn = sline[3]
-
-        # Add one dictionary value per crc
-        if crc not in crcs:
-            crcs[crc] = [pn]
-    f.close()
-
-def make_path(pn):
-    """ Create all folders required to create a full pathname to a folder """
-    # if the folder already exists, we are done without doing a thing
-    if os.path.exists(pn):
-        return
-
-    # Find the parent pathname, abort if that is not possible
-    rootp = os.path.split(pn)
-    if len(rootp) != 2:
-        return
-
-    # Create the parent folder if it does not yet exist
-    if not os.path.exists(rootp[0]):
-        make_path(rootp[0])
-
-    # Create the child folder
-    os.mkdir(pn)
-    return
+    try:
+        os.rename(src, dest)
+        archived += 1
+        print("{} replaced: {}".format(dest, reason))
+    except:
+        print("{} cannot rename to {}".format(src, dest))
+        not_archived += 1
 
 def archive(pn, recursive):
     """ Archive one picture or movie in it's proper place """
     global archive_pictures, picture_exts
-    global archive_videos, movie_exts
-    global months, found, archived, not_archived, removed, crcs
+    global archive_videos, video_exts
+    global months, found, archived, not_archived, removed
 
-    # Optionally process folders recursively
-    if os.path.isdir(pn):
-        for fn in glob.glob(glob.escape(pn) + '\\' + filespec):
-            if os.path.isfile(fn):
-                archive(fn, recursive)
-
-        if recursive:
-            for fn in glob.glob(glob.escape(pn) + '\\*'):
-                if os.path.isdir(fn):
-                    archive(fn, recursive)
-        return
-
-    # only archive some types of files
+    # Only archive some types of files
     rootp, ext = os.path.splitext(pn)
     rootp, fn = os.path.split(pn)
-    if ext.lower() not in picture_exts + movie_exts:
-        print("Not archiving {}".format(pn))
+    if ext.lower() not in picture_exts + video_exts:
+        print("{}: ignoring this type of file".format(pn))
         return
     found = found + 1
-
-    # If a duplicate CRC exists then delete this file
-    crc = crc32(pn)
-    if crc in crcs:
-        if os.path.isfile(crcs[crc]):
-            os.remove(pn)
-            removed += 1
-            print("{} removed as a duplicate".format(pn))
 
     # Decide where to archive the file
     if ext.lower() in picture_exts:
         date = GetExifDate(pn)
         dir_archives = archive_pictures
-    else:
+    elif ext.lower() in video_exts:
+        # fn is correct for this function
         date = GetFileDate(fn)
         dir_archives = archive_videos
+    else:
+        date = None
     if date is None:
-        print("No date found to archive {}".format(pn))
+        print("{}: No archive date".format(pn))
+        not_archived += 1
         return
 
     # Archive pn using computed date
@@ -123,56 +91,56 @@ def archive(pn, recursive):
     month = months[int(date[1])-1]
     #dbg print("The date in {} is {}".format(fn, ymd))
     folder = dir_archives + '\\' + year + '\\' + month
-    make_path(folder)
-    try:
-        backup_pn = folder + '\\' + fn
+    recursive_mkdir(folder)
+
+    # Read the CRCs at the destination.
+    # If a duplicate CRC exists then delete this file
+    crc = crc32(pn)
+    crcs, last_modified = ReadCrcs(folder+"\\crc.csv")
+    backup_pn = folder + '\\' + fn
+    for crcfn in crcs:
+        if crc == crcs[crcfn]:
+            remove_duplicate(pn, folder+'\\'+crcfn)
+            return
+
+    # Try to archive a file simply by renaming it
+    if not os.path.exists(backup_pn):
         os.rename(pn, backup_pn)
+        print("{} archived to {}".format(pn, folder))
         archived = archived + 1
-    except:
-        if crc32(pn) == crc32(backup_pn):
-            os.remove(pn)
-            removed += 1
-            print("{} removed as a duplicate".format(pn))
-        else:
-            # Rename failed and the CRC32 is not the same for the two files
-
-            # Replace the backup files with higher resolution files
-            replace = False
-            src_dim = GetExifDimensions(pn)
-            src_date = GetExifDate(pn)
-            dest_dim = GetExifDimensions(backup_pn)
-            dest_date = GetExifDate(backup_pn)
-            update = False
-            if src_dim is not None:
-                if dest_dim is not None:
-                    if dest_dim[0] < src_dim[0] or dest_dim[1] < src_dim[1]:
-                        update = True   # Improve resolution
-                    else:
-                        os.remove(pn)
-                        removed += 1
-                        print("{} removed as lower resolution".format(pn))
-                        return
-            if src_date is None:
-                if dest_date is not None:
-                    update = True   # Add a date
-
-            if update:
-                os.remove(backup_pn)
-                os.rename(pn, backup_pn)
-                archived = archived + 1
-                print("{} replaced archive in {}".format(pn, folder))
-                return
-
-            print("{} could not be archived to {}".format(pn, folder))
-            not_archived = not_archived + 1
         return
-    print("{} archived to {}".format(pn, folder))
+
+    # Just because the files are named the same yet different does not mean we are done.
+
+    # Replace the backup files with higher resolution files
+    # Remove lower resolution files
+    src_dim = GetExifDimensions(pn)
+    dest_dim = GetExifDimensions(backup_pn)
+    update = False
+    if src_dim is not None and dest_dim is not None:
+        if dest_dim[0] < src_dim[0] and dest_dim[1] < src_dim[1]:
+            replace_destination(pn, backup_pn, "Improve resolution")
+        else:
+            remove_duplicate(pn, backup_pn)
+        return
+
+    # Replace files with no exif date if a date is now available
+    src_date = GetExifDate(pn)
+    dest_date = GetExifDate(backup_pn)
+    if src_date is not None and dest_date is None:
+        replace_destination(pn, backup_pn, "add date")
+        return
+
+    print("{} could not be archived to {}".format(pn, folder))
+    not_archived = not_archived + 1
+    return
 
 def getfolder(folder, filetype):
     """ Pick which folder to use to archive files of a particular file type """
 
-    # The returned value must be on the same drive
-    drive = folder[:2].upper()
+    # The returned value must be on the same drive because rename is used
+    folder = os.path.abspath(folder)
+    drive = folder[0:2].upper()
 
     # Start by looking for an environment variable with the same name
     folder = os.environ.get(filetype)
@@ -191,34 +159,50 @@ def getfolder(folder, filetype):
     # If all else fails, use a folder based upon the root with the same name
     return '\\' + filetype
 
-def main():
-    pass
+def main(pn, recursive):
+    """ Process files and possibly folders starting with a folder """
+
+    # Process just the files in a folder
+    for fn in glob.glob(glob.escape(pn) + '\\' + filespec):
+        if os.path.isfile(fn):
+            archive(fn, recursive)
+
+    # Process folders recursively, when requested
+    if recursive:
+        # Note that the saved filespec cannot be used here
+        for pn in glob.glob(glob.escape(pn) + "\\*"):
+            if os.path.isdir(pn):
+                main(pn, recursive)
+
+def help():
+    print("Photo-Backup [-r] [filespec]")
+    print("\nThis program moves files to pictures and videos folders based")
+    print("upon the date they were recorded.")
+    print("\n-r\trecursively process folders")
+    exit()
 
 if __name__ == '__main__':
-    """ Process command line arguments """
-
     # Process the command line arguments
     folder = os.getcwd();   # <path>: use a path other than the current working directory
     recursive = False       # -r: recursively process subfolders
-    unique_filespec = False # only matters when -r and no folder specified
     for arg in sys.argv[1:]:
-        if arg[0] == '-':
+        if arg[0] in ['-', '/']:
             if arg[1].lower() == 'r':
                 recursive = True
+            else:
+                help()
         elif os.path.isdir(arg):
-            folder = os.path.abspath(arg)
+            folder = arg
         else:
             folder, filespec = os.path.split(arg)
+            folder = os.path.abspath(arg)
 
     # Define the alternative locations to archive pictures and movies
     archive_pictures = getfolder(folder, 'Pictures')
     archive_videos = getfolder(folder, 'Videos')
 
-    # Read in the CRCs of every file previously backed up
-    read_crcs(archive_pictures)
-
-    # Find and process files
-    archive(folder, recursive)
+    # Find and process folders
+    main(folder, recursive)
 
     # Print the result
-    print("Found {}, archived {}, removed {}, {} could not be archived".format(found, archived, removed, not_archived))
+    print("Found {:,}, archived {:,}, removed {:,}, {:,} could not be archived".format(found, archived, removed, not_archived))
